test: separate MOSFET pin mapping from the original repro
